+++ Components/layout/Layout.js
@@ -1,5 +1,10 @@
 import Navigation from "../navigation/Navigation.js";
 
-export default function Layout() {
-  return <Navigation />;
+export default function Layout({ children }) {
+  return (
+    <>
+      <Navigation />
+      <main>{children}</main>
+    </>
+  );
 }
