+++ Components/navigation/Navigation.js
@@ -1,15 +1,15 @@
-import Link from "next/link"
+import Link from "next/link";
 
-export default function Navigation(){
-    return(
-            <ul>
-                <li>
-                    <Link href="/"> ArtPieces <Link/>
-                </li>
-                
-                <li>
-                    <Link href="/"> Spotlight <Link/>
-                </li>
-            </ul>
-    );
+export default function Navigation() {
+  return (
+    <ul>
+      <li>
+        <Link href="/art-pieces"> ArtPieces </Link>
+      </li>
+
+      <li>
+        <Link href="/"> Spotlight </Link>
+      </li>
+    </ul>
+  );
 }
