+++ pages/_app.js
@@ -1,12 +1,31 @@
 import Layout from "@/Components/layout/Layout";
 import GlobalStyle from "../styles";
+import { SWRConfig } from "swr";
 
 export default function App({ Component, pageProps }) {
+  const fetcher = async (url) => {
+    const res = await fetch(url);
+
+    // If the status code is not in the range 200-299,
+    // we still try to parse and throw it.
+    if (!res.ok) {
+      const error = new Error("An error occurred while fetching the data.");
+      // Attach extra info to the error object.
+      error.info = await res.json();
+      error.status = res.status;
+      throw error;
+    }
+
+    return res.json();
+  };
   return (
     <>
       <GlobalStyle />
-      <Component {...pageProps} />
-      <Layout />
+      <Layout>
+        <SWRConfig value={{ fetcher }}>
+          <Component {...pageProps} />
+        </SWRConfig>
+      </Layout>
     </>
   );
 }
