+++ pages/art-pieces/index.js
@@ -1,5 +1,15 @@
 import ArtPieces from "@/Components/artPieces";
+import useSWR from "swr";
 
 export default function ArtPiecesPages() {
-  rendeer(<ArtPieces />);
+  const URL = "https://example-apis.vercel.app/api/art";
+
+  const { data: pieces, error, isLoading } = useSWR(URL);
+  if (error) return <div>failed to load</div>;
+  if (isLoading) return <div>loading...</div>;
+  return (
+    <>
+      <ArtPieces pieces={pieces} />
+    </>
+  );
 }
